+++ lib/dispatch/repositories/request_comments.ex
@@ -16,15 +16,15 @@ defmodule Dispatch.Repositories.RequestComments do
   Given a list of 3 reviewers
 
   %{username: "John", type: "contributor", metadata: %{relevancy: "23/99"}}
-  %{username: "Jane", type: "expert", metadata: %{stack: "elixir"}}
-  %{username: "Joe", type: "expert", metadata: %{stack: "graphql"}}
+  %{username: "Jane", type: "reviewer", metadata: %{stack: "elixir"}}
+  %{username: "Joe", type: "reviewer", metadata: %{stack: "graphql"}}
   %{username: "Jerry", type: "learner"}
 
   So reviewers will be displayed like this:
 
   @John (contributor with 23% relevancy)
-  @Jane (expert reviewer for the elixir stack)
-  @Joe (expert reviewer for the graphql stack)
+  @Jane (reviewer for the elixir stack)
+  @Joe (reviewer for the graphql stack)
   @Jerry (learner reviewer)
   """
   def request_comment(reviewers) do
@@ -49,12 +49,12 @@ defmodule Dispatch.Repositories.RequestComments do
     "* @#{username} (contributor with `#{recent_commit_count}` commits in the last #{relevant_activity_days} days and `#{total_commit_count}` commits overall)\n"
   end
 
-  defp reviewer_line(%{username: username, type: "expert", metadata: %{stack: stack}}) do
-    "* @#{username} (expert reviewer for the `#{stack}` stack)\n"
+  defp reviewer_line(%{username: username, type: "reviewer", metadata: %{stack: stack}}) do
+    "* @#{username} (reviewer for the `#{stack}` stack)\n"
   end
 
   defp reviewer_line(%{username: username, type: "learner", metadata: %{stack: stack}}) do
-    "* @#{username} (learner reviewer for the `#{stack}` stack)\n"
+    "* @#{username} (learner for the `#{stack}` stack)\n"
   end
 
   defp reviewer_line(%{username: username, type: "contributor"}) do
@@ -62,6 +62,6 @@ defmodule Dispatch.Repositories.RequestComments do
   end
 
   defp reviewer_line(%{username: username, type: type}) do
-    "* @#{username} (#{type} reviewer)\n"
+    "* @#{username} (#{type})\n"
   end
 end